feat(docs): extract base doc for package API nav doc
