+++ libs/docs-utils/src/doc/base.type.ts
@@ -1,0 +1,12 @@
+import { DaffBreadcrumb } from '../breadcrumb/public_api';
+import { DaffDocKind } from '../kind/public_api';
+
+/**
+ * A basic generated document that represents some piece of documentation.
+ */
+export interface DaffBaseDoc {
+  id: string;
+  title: string;
+  breadcrumbs: Array<DaffBreadcrumb>;
+  kind: DaffDocKind;
+}
