+++ libs/docs-utils/src/doc/public_api.ts
@@ -2,3 +2,4 @@ export * from './api/public_api';
 export * from './package-guide.type';
 export * from './type';
 export * from './rendered-content.type';
+export * from './base.type';
